Factor out breadboard invocation and set up rendering stub for inputs
